requirements.txt
certifi==2017.11.5
chardet==3.0.4
idna==2.6
Preston==3.2.0
PyYAML==3.12
requests==2.18.4
urllib3==1.22
xmltodict==0.11.0
